requirements.txt
streamlit
pandas
numpy
altair
matplotlib
streamlit-option-menu
textwrap3
pyserial 
pydub